chore: moved edit user form to other place
