+++ src/pages/profile/Dialogs/PhotoDialog.tsx
@@ -1,0 +1,1 @@
+import { Modal } from "../../../components";
